+++ scripts/actors/actors.js
@@ -13,7 +13,6 @@ export function createItemHeaderButton(config, buttons) {
     if (!game.modules.find(m => m.id === "backpack-manager")?.active ?? false) {
         return;
     }
-    debugger;
     if (config.object instanceof Item) {
         var item = config.object;
         if (item.type != "backpack") {
@@ -261,7 +260,6 @@ async function promptForSpellOverrides(actor, overrideableItems, overriddenItems
 }
 
 async function promptForTargetActor(item) {
-    debugger;
 
 
 }